requirements.txt
Flask==2.2.2
redis==4.5.1
python-dotenv==1.0.0
 tiktoken==0.7.0
 Werkzeug==2.3.7